EDG-767: move the browse engine out of the SDK into the framework (one shared engine)
Relocate ProtocolAdapterBrowseEngine + BrowsedNode/BrowseSink/BrowseOutcome to
com.hivemq.protocols.v2.browse (framework, not published SDK API). The opcua
module test scope already depends on the app, so the conformance suite drives
the same single engine as production. Delete the duplicate ReferenceBrowseEngine.
